fix: reduce GL call from CaReactor
